+++ griptape/structures/structure.py
@@ -10,6 +10,7 @@ from attrs import Attribute, Factory, define, field
 from rich.logging import RichHandler
 
 from griptape.artifacts import BaseArtifact, BlobArtifact, TextArtifact
+from griptape.common import observable
 from griptape.config import BaseStructureConfig, OpenAiStructureConfig, StructureConfig
 from griptape.drivers import BaseEmbeddingDriver, BasePromptDriver, OpenAiChatPromptDriver, OpenAiEmbeddingDriver
 from griptape.drivers.vector.local_vector_store_driver import LocalVectorStoreDriver
@@ -255,6 +256,7 @@ class Structure(ABC):
                 if task.id not in child.parent_ids:
                     child.parent_ids.append(task.id)
 
+    @observable
     def before_run(self, args: Any) -> None:
         self._execution_args = args
 
@@ -270,6 +272,7 @@ class Structure(ABC):
 
         self.resolve_relationships()
 
+    @observable
     def after_run(self) -> None:
         self.publish_event(
             FinishStructureRunEvent(
@@ -283,6 +286,7 @@ class Structure(ABC):
     @abstractmethod
     def add_task(self, task: BaseTask) -> BaseTask: ...
 
+    @observable
     def run(self, *args) -> Structure:
         self.before_run(args)